rename section components
