+++ app/hero-section.tsx
